Use dynamic annotation reader and writers
