+++ dlup/__init__.py
@@ -4,7 +4,7 @@ import logging
 
 from ._image import SlideImage
 from ._region import BoundaryMode, RegionView
-from .annotations import AnnotationClass, AnnotationType, WsiAnnotations
+from .annotations import SlideAnnotations
 
 pyvips_logger = logging.getLogger("pyvips")
 pyvips_logger.setLevel(logging.CRITICAL)
@@ -13,4 +13,4 @@ __author__ = """dlup contributors"""
 __email__ = "j.teuwen@nki.nl"
 __version__ = "0.7.0"
 
-__all__ = ("SlideImage", "WsiAnnotations", "AnnotationType", "AnnotationClass", "RegionView", "BoundaryMode")
+__all__ = ("SlideImage", "SlideAnnotations", "RegionView", "BoundaryMode")
